requirements.txt
flask
joblib
pandas
pyyaml
scikit-learn
dvc
